the immutable data load function
